authx - Rename "JWT" API to "AuthxCredential"
